crypto-arb-bot/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.3.0
ccxt==4.5.7
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
propcache==0.3.2
pycares==4.11.0
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
